feat:#14 openai chatGPT 프롬포트 작성
