+++ src/main/java/com/nextPick/questionList/controller/QuestionListController.java
@@ -1,9 +1,9 @@
 package com.nextPick.questionList.controller;
 
 import com.nextPick.questionList.dto.QuestionListDto;
+import com.nextPick.questionList.service.OldQuestionListService;
 import com.nextPick.questionList.service.QuestionListService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.repository.query.Param;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -17,6 +17,7 @@ import javax.validation.constraints.Positive;
 @RequiredArgsConstructor
 @RequestMapping("/questions")
 public class QuestionListController {
+    private final OldQuestionListService oldService;
     private final QuestionListService service;
 
     @PostMapping
@@ -26,10 +27,17 @@ public class QuestionListController {
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    @PostMapping("/{question-id}")
+    @PostMapping("/old/{question-id}")
     public ResponseEntity komoranTest2(@PathVariable("question-id") @Positive long questionId,
             @Valid @RequestBody QuestionListDto.answerCheck answerCheck) {
-        boolean result = service.judgementResponse(questionId, answerCheck.getAnswer());
+        boolean result = oldService.judgementResponse(questionId, answerCheck.getAnswer());
+        return result?new ResponseEntity<>(HttpStatus.OK):new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+    }
+
+    @PostMapping("/{question-id}")
+    public ResponseEntity interviewer(@PathVariable("question-id") @Positive long questionId,
+                                       @Valid @RequestBody QuestionListDto.answerCheck answerCheck) {
+        boolean result = oldService.judgementResponse(questionId, answerCheck.getAnswer());
         return result?new ResponseEntity<>(HttpStatus.OK):new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 }
